2.5.6: Fix \#366 "space in term URI causes malformed SPARQL query"
UriDispatcher now "fixes" each URI prior to instantiating the SPARQL query by replacing spaces with underscores.
Note: although it would probably be more correct to respond with an explicit error, the practical solution
is to just replace each space with an underscore and let the query be processed. This could be documented.
Relevant information:
  - http://stackoverflow.com/a/9056636/830737
  - http://answers.semanticweb.com/questions/8244/how-do-you-handle-non-iri-compatible-urirefs